@@ -1,53 +1,44 @@
 package com.suppergerrie2.panorama;
 
-import static com.suppergerrie2.panorama.Config.panoramaSaveFolder;
-
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.HashMap;
-import java.util.Random;
-import javax.annotation.Nullable;
-import net.minecraft.client.MainWindow;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screen.MainMenuScreen;
-import net.minecraft.client.renderer.RenderSkybox;
-import net.minecraft.client.renderer.RenderSkyboxCube;
+import com.mojang.blaze3d.platform.NativeImage;
+import com.mojang.blaze3d.platform.Window;
+import com.mojang.math.Vector3f;
+import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
+import net.minecraft.client.*;
+import net.minecraft.client.gui.screens.TitleScreen;
+import net.minecraft.client.renderer.CubeMap;
+import net.minecraft.client.renderer.PanoramaRenderer;
 import net.minecraft.client.renderer.texture.DynamicTexture;
-import net.minecraft.client.renderer.texture.NativeImage;
-import net.minecraft.client.settings.KeyBinding;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.ScreenShotHelper;
-import net.minecraft.util.Util;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.math.vector.Vector3f;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.util.text.event.ClickEvent;
-import net.minecraft.util.text.event.ClickEvent.Action;
-import net.minecraftforge.client.event.EntityViewRenderEvent;
-import net.minecraftforge.client.event.GuiOpenEvent;
-import net.minecraftforge.client.event.InputEvent;
-import net.minecraftforge.client.event.RenderWorldLastEvent;
+import net.minecraft.network.chat.ClickEvent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.client.ClientRegistry;
+import net.minecraftforge.client.event.*;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.client.registry.ClientRegistry;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
 
+import javax.annotation.Nullable;
+import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.HashMap;
+import java.util.Random;
+
+import static com.suppergerrie2.panorama.Config.panoramaSaveFolder;
+
 public class PanoramaClientEvents {
 
-    public static final KeyBinding createPanoramaKey = new KeyBinding(
+    public static final KeyMapping createPanoramaKey = new KeyMapping(
             PanoramaMod.MOD_ID + ".key.createPanorama",
             GLFW.GLFW_KEY_H,
             "key.categories." + PanoramaMod.MOD_ID);
-    private static final Logger LOGGER = LogManager.getLogger();
+    private static final Logger     LOGGER            = LogManager.getLogger();
     static HashMap<Path, DynamicTexture[]> skyboxTextureCache = new HashMap<>();
 
     boolean showedWarningMessage = false;
@@ -67,9 +58,9 @@ public class PanoramaClientEvents {
         panoramaSaveFolder = Minecraft.getInstance().gameDirectory.toPath().resolve("panoramas");
     }
 
-    boolean makePanorama = false;
-    Vector3d panoramaPosition = Vector3d.ZERO;
-    long startTime = System.currentTimeMillis();
+    boolean makePanorama     = false;
+    Vec3    panoramaPosition = Vec3.ZERO;
+    long    startTime        = System.currentTimeMillis();
     Vector3f[] stages = new Vector3f[]{
             new Vector3f(0, 0, 0),
             new Vector3f(90, 0, 0),
@@ -82,9 +73,9 @@ public class PanoramaClientEvents {
 
 
     private static void takeScreenshot(final int stage, final long time, boolean sendMessage) {
-        MainWindow window = Minecraft.getInstance().getWindow();
-        final NativeImage screenshot = ScreenShotHelper
-                .takeScreenshot(window.getWidth(), window.getHeight(),
+        Window window = Minecraft.getInstance().getWindow();
+        final NativeImage screenshot = Screenshot
+                .takeScreenshot(/*window.getWidth(), window.getHeight(),*/
                         Minecraft.getInstance().getMainRenderTarget());
 
         Util.ioPool().execute(() -> {
@@ -126,20 +117,20 @@ public class PanoramaClientEvents {
                 if (sendMessage) {
                     File target = path.getParent().toFile();
 
-                    ITextComponent textComponent = (new StringTextComponent(
+                    Component textComponent = (Component.literal(
                             target.getName()))
-                            .withStyle(TextFormatting.UNDERLINE)
+                            .withStyle(ChatFormatting.UNDERLINE)
                             .withStyle((p_238335_1_) ->
                                     p_238335_1_.withClickEvent(
-                                            new ClickEvent(Action.OPEN_FILE,
-                                                    target.getAbsolutePath())));
+                                            new ClickEvent(ClickEvent.Action.OPEN_FILE,
+                                                           target.getAbsolutePath())));
 
                     Minecraft.getInstance().execute(() ->
                             Minecraft.getInstance()
                                     .gui
                                     .getChat()
                                     .addMessage(
-                                            new TranslationTextComponent(
+                                            Component.translatable(
                                                     "spanorama.panorama.success",
                                                     textComponent
                                             )));
@@ -154,7 +145,7 @@ public class PanoramaClientEvents {
                                 .gui
                                 .getChat()
                                 .addMessage(
-                                        new TranslationTextComponent(
+                                        Component.translatable(
                                                 "spanorama.panorama.failed"
                                         )));
             } finally {
@@ -236,36 +227,36 @@ public class PanoramaClientEvents {
     }
 
     /**
-     * Set a random panorama on the given {@link MainMenuScreen}.
+     * Set a random panorama on the given {@link TitleScreen}.
      *
      * @param screen The screen to set the random panorama to, if null only the resources will be
      *               set and not the renderer itself
      */
-    private void setRandomPanorama(@Nullable MainMenuScreen screen) {
+    private void setRandomPanorama(@Nullable TitleScreen screen) {
 
         //If custom panoramas are disabled make sure the vanilla resources are set
         DynamicTexture[] textures = Config.useCustomPanorama ? getRandomPanorama() : null;
-        MainMenuScreen.CUBE_MAP = textures != null ? new RenderDynamicSkyboxCube(
-                textures) : new RenderSkyboxCube(
+        TitleScreen.CUBE_MAP = textures != null ? new RenderDynamicSkyboxCube(
+                textures) : new CubeMap(
                 new ResourceLocation("textures/gui/title/background/panorama"));
         if (screen != null) {
-            screen.panorama = new RenderSkybox(MainMenuScreen.CUBE_MAP);
+            screen.panorama = new PanoramaRenderer(TitleScreen.CUBE_MAP);
         }
     }
 
-    public void openMainMenu(GuiOpenEvent event) {
-        if (event.getGui() instanceof MainMenuScreen) {
+    public void openMainMenu(ScreenOpenEvent event) {
+        if (event.getScreen() instanceof TitleScreen titleScreen) {
             if (!showedWarningMessage && !Config.disableFlashWarning) {
-                event.setGui(new ScreenFlashWarningScreen(event.getGui()));
+                event.setScreen(new ScreenFlashWarningScreen(event.getScreen()));
                 showedWarningMessage = true;
             } else {
-                setRandomPanorama((MainMenuScreen) event.getGui());
+                setRandomPanorama(titleScreen);
             }
         }
     }
 
     @SubscribeEvent
-    void renderEvent(RenderWorldLastEvent event) {
+    void renderEvent(RenderLevelLastEvent event) {
         if (Minecraft.getInstance().level != null && makePanorama) {
             takeScreenshot(stage, startTime, stage == (stages.length - 2));
 
@@ -285,14 +276,14 @@ public class PanoramaClientEvents {
             cameraSetup.setPitch(rotation.y());
             cameraSetup.setRoll(rotation.z());
 
-            cameraSetup.getInfo().setPosition(panoramaPosition);
+            cameraSetup.getCamera().setPosition(panoramaPosition);
         }
     }
 
     @SubscribeEvent
-    void fovModifier(EntityViewRenderEvent.FOVModifier fovModifier) {
+    void fovModifier(EntityViewRenderEvent.FieldOfView fovEvent) {
         if (makePanorama) {
-            fovModifier.setFOV(90);
+            fovEvent.setFOV(90);
         }
     }
 
@@ -311,7 +302,7 @@ public class PanoramaClientEvents {
                 panoramaPosition =
                         Minecraft.getInstance().player != null ?
                                 Minecraft.getInstance().player.getEyePosition(0) :
-                                Vector3d.ZERO;
+                                Vec3.ZERO;
             }
             LOGGER.info("Pressed create panorama key");
         }
